Move inference code to own file
